+++ requirements.txt
@@ -1,2 +1,3 @@
 pygame==2.1.2
 termcolor
+celery
